Ensure that the temperature is rising when it should.
We observerd a strange error yielding in dead tips* we attribute
to shorts between two contacts on the 3.5 mm connector** (bad cable...)
and we hope this failure mode will be covered by this feature.

The feature works by monitoring the temperature and making sure that
in a certain time a certain temperature rising is seen.

The tuning can be done by a few parameters:
  - WATCH_TEMP_PERIOD -- The time the temperature has to rise
  - WATCH_TEMP_INCREASE -- by this value
  - WATCH_TEMP_DEACTIVATE -- if we are that close to the target temperature, disable
        this protection.
  - WATCH_TEMP_REBOUND -- if the temperature drops below the target
        window (e.g by cleaning with a wet sponge), wait this time until
        the protection is re-armed.

* heating up at full power, until red-glowing and then dying)

** it might be tip and ring1, combined with phase reversal on the OpAmp
caused by exceeding the OpAmps Common Mode voltage ratings.. But this is
only a theory...